Moved server files to src, improved TypeScript configuration.
