+++ packages/web-component-tester/.vscode/settings.json
@@ -4,15 +4,6 @@
   "editor.formatOnSave": true,
   "editor.formatOnType": true,
   "files.exclude": {
-    "runner/*.js": true,
-    "runner/*.d.ts": true,
-    "runner/*.js.map": true,
-    "test/unit/*.js": true,
-    "test/unit/*.d.ts": true,
-    "test/unit/*.js.map": true,
-    "test/integration/*.js": true,
-    "test/integration/*.d.ts": true,
-    "test/integration/*.js.map": true,
     "browser.js": true,
     "browser.js.map": true,
     "browser/**/*.js": true,
