+++ packages/web-component-tester/@types/bower-config.d.ts
