--- packages/web-component-tester/mocha.opts
@@ -1,4 +1,0 @@
---ui tdd
---colors
---require source-map-support/register
---timeout 5000
