+++ packages/web-component-tester/package.json
@@ -22,35 +22,36 @@
     "web component",
     "web"
   ],
-  "main": "runner.js",
+  "main": "lib/index.js",
   "bin": {
     "wct": "./bin/wct",
     "wct-st": "./bin/wct-st"
   },
   "files": [
-    "bin/",
-    "data/",
-    "runner/",
-    "scripts/",
-    "tasks/",
     ".bowerrc",
+    "bin",
     "bower.json",
-    "browser.js",
     "browser.js.map",
-    "package.json",
+    "browser.js",
+    "data",
+    "lib",
     "LICENSE",
+    "package.json",
     "README.md",
-    "runner.js"
+    "scripts",
+    "tasks"
   ],
   "scripts": {
     "lint": "gulp lint",
-    "build": "tsc && gulp build",
-    "test": "tsc && gulp test",
-    "prepublishOnly": "gulp prepublish",
-    "test:unit": "mocha \"test/unit/**/*.js\"",
-    "test:integration": "gulp bower && mocha \"test/integration/**/*.js\" --timeout 90000",
-    "test:watch": "watch \"mocha \\\"test/unit/**/*.js\\\"\" runner/ browser/ bin/ test/ tasks/",
-    "format": "find runner test | grep \"\\.js$\\|\\.ts$\" | xargs ./node_modules/.bin/clang-format --style=file -i"
+    "build": "run-p build:*",
+    "build:server": "tsc",
+    "build:browser": "gulp build",
+    "test": "run-p test:*",
+    "prepublishOnly": "run-s test build",
+    "test:unit": "mocha \"test/unit/**/*.ts\"",
+    "test:integration": "gulp bower && mocha \"test/integration/**/*.ts\" --timeout 90000",
+    "test:unit:watch": "watch \"run-s test:unit\" src/ browser/ bin/ test/ tasks/",
+    "format": "find src test | grep \"\\.js$\\|\\.ts$\" | xargs ./node_modules/.bin/clang-format --style=file -i"
   },
   "dependencies": {
     "@polymer/sinonjs": "^1.14.1",
@@ -121,10 +122,12 @@
     "gulp-tslint": "^8.1.2",
     "gulp-typescript": "^3.1.2",
     "lazypipe": "^1.0.1",
+    "npm-run-all": "^4.1.3",
     "rimraf": "^2.5.4",
     "rollup": "^0.25.1",
     "run-sequence": "^1.0.1",
     "source-map-support": "^0.5.4",
+    "ts-node": "^6.0.3",
     "tslint": "^5.9.1",
     "typescript": "^2.1.4",
     "watch": "^0.18.0",
