+++ packages/web-component-tester/src/paths.ts
