+++ packages/web-component-tester/test/mocha.opts
@@ -1,4 +1,5 @@
 --ui tdd
 --colors
 --require source-map-support/register
+--require ts-node/register
 --timeout 5000
